Create dashboard with streamlit
